Add text posts respect the larger preview option, and allow tapping on the preview to navigate to the post instead of having to tap on the title
