+++ app/src/main/res/values/dimens.xml
@@ -14,6 +14,7 @@
 
     <dimen name="post_text_padding">16dp</dimen>
     <dimen name="post_text_max_height">200dp</dimen>
+    <dimen name="post_text_max_height_large">400dp</dimen>
     <dimen name="post_text_card_radius">15dp</dimen>
 
     <dimen name="post_image_height">250dp</dimen>
