+++ app/src/main/res/values/strings.xml
@@ -152,7 +152,7 @@
     <string name="preference_show_nsfw_preview">Show NSFW preview</string>
     <string name="preference_show_spoiler_preview">Show Spoiler preview</string>
     <string name="preference_large_preview">Large previews</string>
-    <string name="preference_large_preview_summary">Double the height of image and video previews in post lists</string>
+    <string name="preference_large_preview_summary">Double the height of image, video and text previews in post lists</string>
 
     <string name="preference_reddit_source">Reddit Source</string>
     <string name="preference_reddit_source_reddit">Reddit</string>
